backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.10.3
moviepy==2.1.1
groq==1.0.0
httpx==0.27.2
python-dotenv==1.0.1
gTTS==2.5.1


